Added VEC_FAST define for optional faster (imprecise) behaviour
VEC_FAST only affects vec_norm (currently) and stores the inverse of the
length of the vector to avoid on unneeded division when (normally)
dividing each vector component by the length. Instead, each vector
component is multiplied by the stored inverse of the length. This often
introduces small floating point inaccuracies which is why this became an
option to begin with.