vec.c: vec_addm: Fixed incorrect freeing of vector.
In the case that vec_addm would be passed a temporary vector as its
first parameter and a permanent vector as its second parameter, vec_addm
would vec_deltemp both vectors, and since the temporary vector would
hold the result, this would mean that the result would be freed.

The needless if else and else causing this issue have therefore been
removed.